Add TryAddCouchbaseBucket extension
Motivation
----------
Need a way to registered a Couchbase bucket which won't register the
bucket again if it is already registered.

Modifications
-------------
Added TryAddCouchbaseBucket to IServiceCollection extensions, which uses
TryAddSingleton instead of AddSingleton to register the bucket.

Results
-------
Consumer can optionally use TryAddCouchbaseBucket, and if the bucket is
already registered it will be ignored.