Attempt to prepend _couchbase._tcp to DNS SRV lookup automatically (#26)
Motivation
----------
The standard for DNS SRV lookup is to lookup using the based domain
name, and then attempt to find the specific service/protocol from that.
So configuring to use "services.local" should lookup the DNS SRV record
for "_couchbase._tcp.services.local" automatically.  This is consistent
with the behavior of other Couchbase SDKs.

Modifications
-------------
Attempt to prepend _couchbase._tcp automatically (so long as the domain
doesn't already start with an underscore).  For backwards compatibility,
fallback to checking the provided record without prepending anything if
no records are found.

Results
-------
Behavior is now more consistent with DNS standards and other Couchbase
SDKs.