DNS SRV support for connection strings
Motivation
----------
Support the DNS SRV portion of the couchbase:// connection string
specification as transparently as possible.

Modifications
-------------
Upgrade Couchbase SDK to 2.5.9, which includes couchbase:// connection
string support.

Detect the use of the connection string, and when compatible
automatically perform DNS SRV lookups.  Replace the connection string
with the list of servers, if found.

Ignore the priority on the DNS SRV entries to be consitent with the
specification.

Results
-------
So long as .AddCouchbaseDnsDiscovery() is applied to the DI container,
DNS SRV discovery following the Couchbase specification
https://github.com/couchbaselabs/sdk-rfcs/blob/master/rfc/0011-connection-string.md#srv-records
is performed automatically.

If using the legacy .AddCouchbaseDnsDiscovery(recordName) approach,
priority is still applied to maintain backwards compatibility.