Upgrade locks to SDK 3.x
Motivation
----------
Provide support for distributed mutexes to users of the 3.x SDK.

Modifications
-------------
Upgrade to SDK 3.x. Rewrite mutexes to be based on an
ICouchbaseCollection instead of an IBucket, to use exceptions instead of
responses containing errors, and the new logging infrastructure.

Upgrade to C# 8 and enable nullable reference types.

Update unit tests, the example project, and documentation.

Results
-------
SDK 3.x compatibility.

Closes #83